// src/modules/models/dto-teacher.ts
export interface TeacherDTO {
  id?: string;
  teacher_name:       string
  teacher_email:      string
  teacher_password:   string
  teacher_phone:      string
  teacher_register:   string
  teacher_education:  string
  teacher_experience: string
  teacher_description: string
  teacher_gender:     string
  teacher_image:      string
  teacher_status:     number
  teacher_birthday:   string
}